+++ pyscript.core/tests/python/index.html
@@ -4,8 +4,8 @@
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>Pure Python PyScript tests</title>
-        <link rel="stylesheet" href="../../../dist/core.css">
-        <script type="module" src="../../../dist/core.js"></script>
+        <link rel="stylesheet" href="../../dist/core.css">
+        <script type="module" src="../../dist/core.js"></script>
     </head>
     <body>
         <script src="./helper.js"></script>
