+++ pyscript.core/tests/python/main.py
@@ -3,6 +3,6 @@ import json
 import upytest
 from pyscript import web
 
-result = await upytest.run("./tests")
+result = await upytest.run("./tests", random=True)
 output = web.div(json.dumps(result), id="result")
 web.page.append(output)
