+++ pyscript.core/tests/python/settings_py.json
@@ -1,0 +1,27 @@
+{
+    "files": {
+        "https://raw.githubusercontent.com/ntoll/upytest/1.0.8/upytest.py": "",
+        "./tests/test_config.py": "tests/test_config.py",
+        "./tests/test_current_target.py": "tests/test_current_target.py",
+        "./tests/test_display.py": "tests/test_display.py",
+        "./tests/test_document.py": "tests/test_document.py",
+        "./tests/test_fetch.py": "tests/test_fetch.py",
+        "./tests/test_ffi.py": "tests/test_ffi.py",
+        "./tests/test_js_modules.py": "tests/test_js_modules.py",
+        "./tests/test_storage.py": "tests/test_storage.py",
+        "./tests/test_running_in_worker.py": "tests/test_running_in_worker.py",
+        "./tests/test_web.py": "tests/test_web.py",
+        "./tests/test_websocket.py": "tests/test_websocket.py",
+        "./tests/test_when.py": "tests/test_when.py",
+        "./tests/test_window.py": "tests/test_window.py"
+    },
+    "js_modules": {
+        "main": {
+            "./example_js_module.js": "greeting"
+        },
+        "worker": {
+            "./example_js_worker_module.js": "greeting_worker"
+        }
+    },
+    "packages": ["Pillow" ]
+}
